pyproject.toml
[project]
name = "text_mate_backend"
version = "0.1.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dcc-backend-common>=0.0.3",
    "dependency-injector>=4.48.3",
    "fastapi-azure-auth>=5.2.0",
    "fastapi[standard]>=0.128.0",
    "pydantic-ai>=1.44.0",
    "pydantic-ai-slim[openai]>=1.44.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.5",
    "returns>=0.26.0",
    "structlog>=25.5.0",
    "tqdm>=4.67.1",
    "types-requests>=2.32.4.20260107",
    "types-setuptools>=80.9.0.20251223",
    "types-tqdm>=4.67.0.20250809",
]

[dependency-groups]
dev = [
    "coverage>=7.6.12",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.22.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.13",
    "version-pioneer>=0.0.13",
    "ty>=0.0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/text_mate_backend"]

[tool.ruff]
# Same as Black.
line-length = 120
indent-width = 4

target-version = "py313"

[tool.ruff.lint]
# Add rules to enable
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
]

# Ignore specific rules
ignore = []

[tool.ruff.format]
# Use double quotes for strings.
quote-style = "double"

# Indent with spaces, rather than tabs.
indent-style = "space"

# Respect magic trailing commas.
skip-magic-trailing-comma = false

# Automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint.isort]
# Group imports by type (standard library, third-party, first-party).
combine-as-imports = true
force-wrap-aliases = true



[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--import-mode=importlib",
]
pythonpath = ["src"]
